RF+PEP8: Move the OpenFMRI tests into their own file (with PEP8 improvements)
