+++ mvpa2/tests/test_openfmri.py
@@ -1,0 +1,96 @@
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+#
+#   See COPYING file distributed along with the PyMVPA package for the
+#   copyright and license terms.
+#
+### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+"""Unit tests for PyMVPA's OpenFMRI data source adaptor"""
+
+import os
+import numpy as np
+
+from mvpa2.testing import *
+
+if not externals.exists('nibabel'):
+    raise SkipTest
+
+from mvpa2 import pymvpa_dataroot
+from mvpa2.datasets.sources.openfmri import OpenFMRIDataset
+from mvpa2.misc.io.base import SampleAttributes
+
+
+def test_openfmri_dataset():
+    of = OpenFMRIDataset(os.path.join(pymvpa_dataroot, 'openfmri'))
+    sub_ids = of.get_subj_ids()
+    assert_equal(sub_ids, [1, 'phantom'])
+    assert_equal(of.get_scan_properties(), {'TR': '2.5'})
+    tasks = of.get_task_descriptions()
+    assert_equal(tasks, {1: 'object viewing'})
+    task = tasks.keys()[0]
+    run_ids = of.get_bold_run_ids(sub_ids[0], task)
+    assert_equal(run_ids, range(1, 13))
+    task_runs = of.get_task_bold_run_ids(task)
+    assert_equal(task_runs, {1: range(1, 13)})
+
+    orig_attrs = SampleAttributes(os.path.join(pymvpa_dataroot,
+                                               'attributes_literal.txt'))
+    for subj, runs in task_runs.iteritems():
+        for run in runs:
+            # load single run
+            ds = of.get_bold_run_dataset(subj, task, run, flavor='1slice',
+                                         mask=os.path.join(pymvpa_dataroot,
+                                                           'mask.nii.gz'),
+                                         add_sa='bold_moest.txt')
+            # basic shape
+            assert_equal(len(ds), 121)
+            assert_equal(ds.nfeatures, 530)
+            # functional mapper
+            assert_equal(ds.O.shape, (121, 40, 20, 1))
+            # additional attributes present
+            moest = of.get_bold_run_motion_estimates(subj, task, run)
+            for i in range(6):
+                moest_attr = 'bold_moest.txt_%i' % (i,)
+                assert_true(moest_attr in ds.sa)
+                assert_array_equal(moest[:, i], ds.sa[moest_attr].value)
+
+            # check conversion of model into sample attribute
+            events = of.get_bold_run_model(subj, task, run)
+            for i, ev in enumerate(events):
+                # we only have one trial per condition in the demo dataset
+                assert_equal(ev['conset_idx'], 0)
+                # proper enumeration and events sorted by time
+                assert_equal(ev['onset_idx'], i)
+            onsets = [e['onset'] for e in events]
+            sorted_onsets = sorted(onsets)
+            assert_array_equal(sorted_onsets, onsets)
+
+            targets = events2sample_attr(events,
+                                         ds.sa.time_coords,
+                                         noinfolabel='rest')
+            assert_array_equal(
+                orig_attrs['targets'][(run - 1) * 121: run * len(ds)], targets)
+            assert_equal(ds.sa['subj'][0], subj)
+
+            # check that we can get the same result from the model dataset
+            # (make it exercise the preproc interface too)
+
+            def preproc_img(img):
+                return img
+            modelds = of.get_model_bold_dataset(
+                1, subj, flavor='1slice',
+                preproc_img=preproc_img,
+                modelfx=assign_conditionlabels,
+                mask=os.path.join(pymvpa_dataroot,
+                                  'mask.nii.gz'),
+                add_sa='bold_moest.txt')
+            modelds = modelds[modelds.sa.run == run]
+            targets = np.array(targets, dtype='object')
+            targets[targets == 'rest'] = None
+            assert_array_equal(targets, modelds.sa.targets)
+    # more basic access
+    motion = of.get_task_bold_attributes(1, 'bold_moest.txt', np.loadtxt)
+    assert_equal(len(motion), 12)  # one per run
+    # one per subject, per volume, 6 estimates
+    assert_equal([m.shape for m in motion], [(1, 121, 6)] * 12)
